refactor: moved data table categories
